Meilleur lisibilité
moins de saut de ligne
Correction bug dans synthèse quotidienne.

+++ src/main/java/org/djer/calendarCalculator/model/CalendarCalculator.java
@@ -38,7 +38,7 @@ public class CalendarCalculator {
     }
 
     public static void addStringSumary(StringBuilder sb, int currentDay, int nbEventThisDay,
-            int cumulativeDurationPerDay) {
+            long cumulativeDurationPerDay) {
 
         sb.append("======================= ");
         sb.append(" Jour : ");
